🕵Review Feedback (#569)
* refactor: make GroupLoader ABC

* refactor: use composition for clean command

* refactor: renamed
