+++ cognite_toolkit/_cdf_tk/commands/clean.py
@@ -42,8 +42,10 @@ from cognite_toolkit._cdf_tk.utils import (
     read_yaml_file,
 )
 
+from ._utils import _print_ids_or_length, _remove_duplicates
 
-class CleanBaseCommand(ToolkitCommand):
+
+class CleanCommand(ToolkitCommand):
     def clean_resources(
         self,
         loader: ResourceLoader,
@@ -78,7 +80,7 @@ class CleanBaseCommand(ToolkitCommand):
         # Duplicates should be handled on the build step,
         # but in case any of them slip through, we do it here as well to
         # avoid an error.
-        loaded_resources, duplicates = self._remove_duplicates(loaded_resources, loader)
+        loaded_resources, duplicates = _remove_duplicates(loaded_resources, loader)
 
         capabilities = loader.get_required_capability(loaded_resources)
         if capabilities:
@@ -133,25 +135,25 @@ class CleanBaseCommand(ToolkitCommand):
         if dry_run:
             nr_of_deleted += len(resource_ids)
             if verbose:
-                print(f"  Would have deleted {self._print_ids_or_length(resource_ids)}.")
+                print(f"  Would have deleted {_print_ids_or_length(resource_ids)}.")
             return nr_of_deleted
 
         try:
             nr_of_deleted += loader.delete(resource_ids)
         except CogniteAPIError as e:
-            self.warn(MediumSeverityWarning(f"Failed to delete {self._print_ids_or_length(resource_ids)}. Error {e}."))
+            self.warn(MediumSeverityWarning(f"Failed to delete {_print_ids_or_length(resource_ids)}. Error {e}."))
             if verbose:
                 print(Panel(traceback.format_exc()))
         except CogniteNotFoundError:
             if verbose:
-                print(f"  [bold]INFO:[/] {self._print_ids_or_length(resource_ids)} do(es) not exist.")
+                print(f"  [bold]INFO:[/] {_print_ids_or_length(resource_ids)} do(es) not exist.")
         except Exception as e:
-            self.warn(MediumSeverityWarning(f"Failed to delete {self._print_ids_or_length(resource_ids)}. Error {e}."))
+            self.warn(MediumSeverityWarning(f"Failed to delete {_print_ids_or_length(resource_ids)}. Error {e}."))
             if verbose:
                 print(Panel(traceback.format_exc()))
         else:  # Delete succeeded
             if verbose:
-                print(f"  Deleted {self._print_ids_or_length(resource_ids)}.")
+                print(f"  Deleted {_print_ids_or_length(resource_ids)}.")
         return nr_of_deleted
 
     def _drop_data(
@@ -194,84 +196,21 @@ class CleanBaseCommand(ToolkitCommand):
         if drop_count > 0:
             print(
                 f"  {prefix} {drop_count:,} {loader.item_name} from {loader.display_name}: "
-                f"{self._print_ids_or_length(resource_ids)}."
+                f"{_print_ids_or_length(resource_ids)}."
             )
         elif drop_count == 0:
             verb = "is" if len(resource_ids) == 1 else "are"
             print(
-                f"  The {loader.display_name}: {self._print_ids_or_length(resource_ids)} {verb} empty, "
+                f"  The {loader.display_name}: {_print_ids_or_length(resource_ids)} {verb} empty, "
                 f"thus no {loader.item_name} will be {'touched' if dry_run else 'dropped'}."
             )
         else:
             # Count is not supported
             print(
                 f" {prefix} all {loader.item_name} from {loader.display_name}: "
-                f"{self._print_ids_or_length(resource_ids)}."
+                f"{_print_ids_or_length(resource_ids)}."
             )
 
-    def _load_files(
-        self,
-        loader: ResourceLoader,
-        filepaths: list[Path],
-        ToolGlobals: CDFToolConfig,
-        skip_validation: bool,
-        verbose: bool = False,
-    ) -> T_CogniteResourceList | None:
-        loaded_resources = loader.create_empty_of(loader.list_write_cls([]))
-        for filepath in filepaths:
-            try:
-                resource = loader.load_resource(filepath, ToolGlobals, skip_validation)
-            except KeyError as e:
-                # KeyError means that we are missing a required field in the yaml file.
-                print(
-                    f"[bold red]ERROR:[/] Failed to load {filepath.name} with {loader.display_name}. Missing required field: {e}."
-                    f"[bold red]ERROR:[/] Please compare with the API specification at {loader.doc_url()}."
-                )
-                return None
-            except Exception as e:
-                print(f"[bold red]ERROR:[/] Failed to load {filepath.name} with {loader.display_name}. Error: {e!r}.")
-                if verbose:
-                    print(Panel(traceback.format_exc()))
-                return None
-            if resource is None:
-                # This is intentional. It is, for example, used by the AuthLoader to skip groups with resource scopes.
-                continue
-            if isinstance(resource, loader.list_write_cls) and not resource:
-                self.warn(LowSeverityWarning(f"Skipping {filepath.name}. No data to load."))
-                continue
-
-            if isinstance(resource, loader.list_write_cls):
-                loaded_resources.extend(resource)
-            else:
-                loaded_resources.append(resource)
-        return loaded_resources
-
-    @staticmethod
-    def _print_ids_or_length(resource_ids: SequenceNotStr[T_ID], limit: int = 10) -> str:
-        if len(resource_ids) == 1:
-            return f"{resource_ids[0]!r}"
-        elif len(resource_ids) <= limit:
-            return f"{resource_ids}"
-        else:
-            return f"{len(resource_ids)} items"
-
-    def _remove_duplicates(
-        self, loaded_resources: T_CogniteResourceList, loader: ResourceLoader
-    ) -> tuple[T_CogniteResourceList, list[T_ID]]:
-        seen: set[T_ID] = set()
-        output = loader.create_empty_of(loaded_resources)
-        duplicates: list[T_ID] = []
-        for item in loaded_resources:
-            identifier = loader.get_id(item)
-            if identifier not in seen:
-                output.append(item)
-                seen.add(identifier)
-            else:
-                duplicates.append(identifier)
-        return output, duplicates
-
-
-class CleanCommand(CleanBaseCommand):
     def execute(
         self, ctx: typer.Context, build_dir_raw: str, build_env_name: str, dry_run: bool, include: list[str]
     ) -> None:
